@@ -9,6 +9,7 @@
     "update-schema": "babel-node ./scripts/updateSchema.js"
   },
   "dependencies": {
+    "accounting": "^0.4.1",
     "babel-core": "^6.5.2",
     "babel-loader": "6.2.3",
     "babel-polyfill": "6.5.0",
@@ -28,8 +29,8 @@
     "morgan": "^1.7.0",
     "pg": "^4.5.2",
     "pg-hstore": "^2.3.2",
-    "react": "0.14.7",
-    "react-dom": "0.14.7",
+    "react": "^15.0.0-rc.2",
+    "react-dom": "^15.0.0-rc.2",
     "react-relay": "0.7.3",
     "react-router": "^2.0.1",
     "react-router-relay": "^0.9.5",
