@@ -7,15 +7,24 @@ import {
 } from '../components'
 
 const App = (props) => {
-  const { buildingComponent, unitComponent, company } = props
+  const {
+    company,
+    buildingComponent,
+    propertyComponent,
+    unitComponent,
+    unitDetailComponent,
+  } = props
   console.log('render App - props:', props)
   return (
     <div className='ui four column grid'>
       <div className={cx('segments', 'ui', 'row attached segment')}>
-        <PropertyListContainer company={company} />
+        {propertyComponent}
         {buildingComponent || <noscript />}
         {unitComponent || <noscript />}
         <div className='ui top attached violet label'>{company.name}</div>
+      </div>
+      <div className='row'>
+        {unitDetailComponent || <noscript />}
       </div>
     </div>
   )
@@ -23,8 +32,10 @@ const App = (props) => {
 
 App.propTypes = {
   company: PropTypes.object.isRequired,
+  propertyComponent: PropTypes.object.isRequired,
   buildingComponent: PropTypes.object,
   unitComponent: PropTypes.object,
+  unitDetailComponent: PropTypes.object,
 }
 
 export default Relay.createContainer(App, {
