@@ -4,8 +4,10 @@ import Relay from 'react-relay'
 import { Route } from 'react-router'
 
 import {
+  PropertyListContainer,
   BuildingListContainer,
   UnitListContainer,
+  UnitDetailContainer,
   NotFound,
 } from './components'
 
@@ -28,24 +30,53 @@ const UnitQueries = {
     query UnitQueries { node(id: $buildingId) }
   `,
 }
+const UnitDetailQueries = {
+  unit: () => Relay.QL`
+    query UnitDetailQueries { node(id: $unitId) }
+  `
+}
 const routes = [
   (<Route
     key='route1'
-    path='/'
-    component={AppContainer}
-    queries={PropertyQueries}>
+    queries={PropertyQueries}
+    component={AppContainer}>
     <Route
-      path=':propertyId/buildings'
-      components={{ buildingComponent: BuildingListContainer }}
-      queries={{ buildingComponent: BuildingQueries }} />
-    <Route path=':propertyId/buildings/:buildingId/units'
+      path='/'
+      components={{ propertyComponent: PropertyListContainer }}
+      queries={{ propertyComponent: PropertyQueries }} />
+    <Route
+      path='/:propertyId/buildings'
       components={{
+        propertyComponent: PropertyListContainer,
+        buildingComponent: BuildingListContainer,
+      }}
+      queries={{
+        propertyComponent: PropertyQueries,
+        buildingComponent: BuildingQueries,
+      }} />
+    <Route path='/:propertyId/buildings/:buildingId/units'
+      components={{
+        propertyComponent: PropertyListContainer,
         buildingComponent: BuildingListContainer,
         unitComponent: UnitListContainer,
       }}
       queries={{
+        propertyComponent: PropertyQueries,
         buildingComponent: BuildingQueries,
         unitComponent: UnitQueries }} />
+    <Route path='/:propertyId/buildings/:buildingId/units/:unitId'
+      components={{
+        propertyComponent: PropertyListContainer,
+        buildingComponent: BuildingListContainer,
+        unitComponent: UnitListContainer,
+        unitDetailComponent: UnitDetailContainer,
+      }}
+      queries={{
+        propertyComponent: PropertyQueries,
+        buildingComponent: BuildingQueries,
+        unitComponent: UnitQueries,
+        unitDetailComponent: UnitDetailQueries,
+      }} />
   </Route>),
   <Route key='route2' path='*' component={NotFound} />,
 ]
